fix:small ui issues
